# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ApprovalDecisionRequest": {
        "properties": {
          "reason": {
            "maxLength": 4000,
            "minLength": 3,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "ApprovalDecisionRequest",
        "type": "object"
      },
      "ApprovalStatus": {
        "enum": [
          "PENDING",
          "APPROVED",
          "REJECTED",
          "EXPIRED",
          "CANCELLED"
        ],
        "title": "ApprovalStatus",
        "type": "string"
      },
      "ApprovalView": {
        "properties": {
          "approver_constraints": {
            "additionalProperties": true,
            "title": "Approver Constraints",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "decided_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          },
          "decided_by": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided By"
          },
          "decision_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Reason"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "policy_decision_id": {
            "format": "uuid",
            "title": "Policy Decision Id",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "requested_by": {
            "format": "uuid",
            "title": "Requested By",
            "type": "string"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ApprovalStatus"
          },
          "step_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Step Id"
          }
        },
        "required": [
          "id",
          "run_id",
          "step_id",
          "policy_decision_id",
          "status",
          "reason",
          "requested_by",
          "approver_constraints",
          "decided_by",
          "decision_reason",
          "expires_at",
          "decided_at",
          "created_at"
        ],
        "title": "ApprovalView",
        "type": "object"
      },
      "ArtifactKind": {
        "enum": [
          "TEXT",
          "TABLE",
          "CHART",
          "SQL",
          "CODE",
          "DIFF",
          "REPORT",
          "DASHBOARD",
          "FILE",
          "DIAGRAM"
        ],
        "title": "ArtifactKind",
        "type": "string"
      },
      "ArtifactView": {
        "properties": {
          "classification": {
            "$ref": "#/components/schemas/Classification"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "inline_content": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inline Content"
          },
          "kind": {
            "$ref": "#/components/schemas/ArtifactKind"
          },
          "lineage": {
            "additionalProperties": true,
            "title": "Lineage",
            "type": "object"
          },
          "media_type": {
            "title": "Media Type",
            "type": "string"
          },
          "run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "storage_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Storage Key"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "workspace_id": {
            "format": "uuid",
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "workspace_id",
          "run_id",
          "kind",
          "title",
          "media_type",
          "inline_content",
          "storage_key",
          "classification",
          "lineage",
          "created_at"
        ],
        "title": "ArtifactView",
        "type": "object"
      },
...
      "Body_upload_artifact_api_v1_workspaces__workspace_id__artifacts_post": {
        "properties": {
          "classification": {
            "$ref": "#/components/schemas/Classification",
            "default": "INTERNAL"
          },
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/ArtifactKind",
            "default": "FILE"
          },
          "lineage": {
            "default": "{}",
            "maxLength": 20000,
            "title": "Lineage",
            "type": "string"
          },
          "run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "title": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "file",
          "title"
        ],
        "title": "Body_upload_artifact_api_v1_workspaces__workspace_id__artifacts_post",
        "type": "object"
      },
...
      "Classification": {
        "enum": [
          "PUBLIC",
          "INTERNAL",
          "CONFIDENTIAL",
          "RESTRICTED"
        ],
        "title": "Classification",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "summary": "Enterprise Agent Runtime and Intelligence Workspace",
    "title": "Obsion API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/approvals": {
      "get": {
        "operationId": "list_approvals_api_v1_approvals_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ApprovalStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ApprovalView"
                  },
                  "title": "Response List Approvals Api V1 Approvals Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Approvals",
        "tags": [
          "approvals"
        ]
      }
    },
    "/api/v1/approvals/{approval_id}/approve": {
      "post": {
        "operationId": "approve_api_v1_approvals__approval_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "approval_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Approval Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Approve",
        "tags": [
          "approvals"
        ]
      }
    },
    "/api/v1/approvals/{approval_id}/reject": {
      "post": {
        "operationId": "reject_api_v1_approvals__approval_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "approval_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Approval Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reject",
        "tags": [
          "approvals"
        ]
      }
    },
    "/api/v1/artifacts/{artifact_id}": {
      "get": {
        "operationId": "get_artifact_api_v1_artifacts__artifact_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Artifact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Artifact",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v1/artifacts/{artifact_id}/content": {
      "get": {
        "operationId": "download_artifact_api_v1_artifacts__artifact_id__content_get",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Artifact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Download Artifact",
        "tags": [
          "artifacts"
        ]
      }
    },
...
    "/api/v1/workspaces/{workspace_id}/artifacts": {
      "get": {
        "operationId": "list_workspace_artifacts_api_v1_workspaces__workspace_id__artifacts_get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ArtifactView"
                  },
                  "title": "Response List Workspace Artifacts Api V1 Workspaces  Workspace Id  Artifacts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Workspace Artifacts",
        "tags": [
          "artifacts"
        ]
      },
      "post": {
        "operationId": "upload_artifact_api_v1_workspaces__workspace_id__artifacts_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_artifact_api_v1_workspaces__workspace_id__artifacts_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload Artifact",
        "tags": [
          "artifacts"
        ]
      }
    },
...
  }
}